refactor: split sandbox cli config
