+++ services/api-rs/crates/centaur-api-server/src/main.rs
@@ -3,7 +3,7 @@ mod cli;
 use centaur_api_server::build_router_with_runtime;
 use centaur_session_sqlx::PgSessionStore;
 use clap::Parser;
-use cli::{Cli, ServerError, sandbox_runtime_from_args};
+use cli::{Cli, ServerError};
 use tokio::net::TcpListener;
 use tracing::info;
 use tracing_subscriber::{EnvFilter, fmt as tracing_fmt};
@@ -19,7 +19,7 @@ async fn main() -> Result<(), ServerError> {
     if cli.run_migrations {
         store.run_migrations().await?;
     }
-    let sandbox_runtime = sandbox_runtime_from_args(&cli.sandbox).await?;
+    let sandbox_runtime = cli.sandbox.runtime().await?;
 
     let listener = TcpListener::bind(cli.bind_addr).await?;
     info!(bind_addr = %cli.bind_addr, "starting centaur api-rs server");
